pyproject.toml
[tool.poetry]
name = "wikitable_clean"
version = "0.1.0"
description = ""
authors = ["Clapika"]

[tool.poetry.dependencies]
python = "^3.8"
transformers = "^4.19.2"
jsonlines = "^3.0.0"
pandas = "^1.4.2"
datasets = "^2.2.2"
spacy = "^3.3.0"
blingfire = "^0.1.8"
elasticsearch-dsl = "^7.4.0"
scikit-learn = "^1.1.1"
lxml = "^4.9.0"
wikipedia = "^1.4.0"
accelerate = "^0.9.0"
orjson = "^3.7.2"
inflect = "^5.6.0"
pyserini = "^0.17.0"
wandb = "^0.12.20"
wikitextparser = "^0.49.2"
flair = "^0.11.3"
dateparser = "^1.1.1"
price-parser = "^0.3.4"
number-parser = "^0.2.1"
nltk = "^3.7"
pytorch-lightning = "^1.6.4"
torchvision = "0.12.0+cu113"
sentence-transformers = "^2.2.2"
gradio = "^3.0.24"
colorama = "^0.4.5"
mwparserfromhell = "^0.6.4"


[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu113/"
secondary = true